+++ CHANGES.md
@@ -10,6 +10,7 @@
 
 ## Fixes
 
+- Suppress signature help for completed applications. (#1982, #1162, @rgrinberg)
 - Exclude surrounding parentheses from infix operator rename placeholders and
   edits. (#1966, fixes #1190, @rgrinberg)
 - Suppress signature help for applications after the cursor. (#1965, #1162,